New Readme file
an application that will provide an easy way for the people to workout on a daily basis. This application will help user to organize their exercise on the basis of days by providing a feature to checklist their exercise that has been done by them and adding comments on the basis of performance.The comments will be saved for future use and could be deleted once the goal is achieved. In overall, our application will provide user a tool to add their exercise, marking them once it is done and providing them a feature to add and delete comments.
Primary Goals
Creating an application using a tableview as a preference in the starting, which will display the days of the week.
Providing user a feature to add names(workouts) depending on the user preference for the selected day.
Also have a feature to check their exercise once it is done using a tableview as a preference in the starting.
Allows the user to add comments on the basis of their performance.
Store the workout list and comments for user’s future use for a particular day.
Pipedream Goals       
We would like to add another checklist for diet control.
Interface our application to push notifications to remind users without being in the app.